Move vespa search tests under dedicated vespa folder
